requirements.txt
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
pyyaml>=6.0
tqdm>=4.65.0
mlflow>=2.10.0
tensorboard>=2.15.0
